fix: jspsych data properties moved to app
