+++ public/electron.js
@@ -204,9 +204,6 @@ ipc.on('data', (event, args) => {
 
     //write the data
     stream.write(JSON.stringify({...args, git}))
-
-    // Copy provocation images to participant's data folder
-    if (args.trial_type === 'image_keyboard_response') images.push(args.stimulus.slice(7))
   }
 })
