+++ src/components/JsPsychExperiment.js
@@ -1,13 +1,9 @@
 import React from "react";
 import { Experiment, jsPsych } from "jspsych-react";
 import { tl } from "../timelines/main";
-import "../App.css";
-import "bootstrap/dist/css/bootstrap.css";
-import "@fortawesome/fontawesome-free/css/all.css";
 import { rt_categorize_html } from "../lib/rt-categorize-html";
 
 function JsPsychExperiment({ dataUpdateFunction, dataFinishFunction }) {
-  console.log("Outside Turk:", jsPsych.turk.turkInfo().outsideTurk);
   jsPsych.plugins["rt-categorize-html"] = rt_categorize_html();
 
   return (
